securing task access

+++ server/taskRepository.js
@@ -2,7 +2,36 @@ import postgres from "postgres";
 
 const sql = postgres();
 
-const create = async (todoId, task) => {
+const isTaskOwnedByUser = async (userId, taskId) => {
+  const userTodoTaskCount = await sql`
+    SELECT COUNT(*) AS count
+      FROM todo_tasks
+      JOIN todos ON todo_tasks.todo_id = todos.id
+      WHERE todo_tasks.id = ${taskId} AND todos.user_id = ${userId}`;
+
+  if (userTodoTaskCount[0]?.count === "0") {
+    return false;
+  }
+  return true;
+};
+
+const isTodoOwnedByUser = async (userId, todoId) => {
+  const userTodoCount = await sql`
+    SELECT COUNT(*) AS count
+      FROM todos
+      WHERE todos.id = ${todoId} AND todos.user_id = ${userId}`;
+
+  if (userTodoCount[0]?.count === "0") {
+    return false;
+  }
+  return true;
+};
+
+const create = async (userId, todoId, task) => {
+  if (!await isTodoOwnedByUser(userId,todoId)) {
+    return null;
+  }
+
   const result = await sql`INSERT INTO todo_tasks
     (todo_id, description)
     VALUES (${todoId}, ${task.description})
@@ -11,27 +40,46 @@ const create = async (todoId, task) => {
   return result[0];
 };
 
-const findAll = async (todoId) => {
+const findAll = async (userId, todoId) => {
+  if (!await isTodoOwnedByUser(userId,todoId)) {
+    return null;
+  }
+
   return await sql`SELECT * FROM todo_tasks
   WHERE todo_id = ${todoId};`;
 };
 
-const findById = async (id) => {
+const findById = async (userId, id) => {
+  if (!await isTaskOwnedByUser(userId, id)) {
+    return null;
+  }
+
   const result = await sql`SELECT * FROM todo_tasks
   WHERE id = ${id}`;
   return result[0];
 };
 
-const updateById = async (id, task) => {
-  const result = await sql`UPDATE todo_tasks SET
+const updateById = async (userId, id, task) => {
+  if (!await isTaskOwnedByUser(userId, id)) {
+    return null;
+  }
+
+  const result = await sql`
+    UPDATE todo_tasks
+    SET
       description = ${task.description},
       is_done = ${task.is_done}
     WHERE id = ${id}
-    RETURNING *;`;
+    RETURNING *`;
+
   return result[0];
 };
 
-const deleteById = async (id) => {
+const deleteById = async (userId, id) => {
+  if (!await isTaskOwnedByUser(userId,id)) {
+    return null;
+  }
+
   const result = await sql`DELETE FROM todo_tasks
     WHERE id = ${id} RETURNING *`;
   return result[0];
